requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3 